IA/requirements.txt
django
djangorestframework
transformers
torch
pillow
